refactor s to arrow functions
