+++ frontend/src/lib/emotion.ts
@@ -1,11 +1,11 @@
-import createCache from "@emotion/cache";
+import createCache, { EmotionCache } from "@emotion/cache";
 
 const isBrowser = typeof document !== "undefined";
 
 // On the client side, Create a meta tag at the top of the <head> and set it as insertionPoint.
 // This assures that MUI styles are loaded first.
 // It allows developers to easily override MUI styles with other styling solutions, like CSS modules.
-export function createEmotionCache() {
+export const createEmotionCache = (): EmotionCache => {
   let insertionPoint;
 
   if (isBrowser) {
@@ -14,4 +14,4 @@ export function createEmotionCache() {
   }
 
   return createCache({ key: "mui-style", insertionPoint });
-}
+};
